python/ucxx/pyproject.toml
# SPDX-FileCopyrightText: Copyright (c) 2023-2026, NVIDIA CORPORATION & AFFILIATES.
# SPDX-License-Identifier: BSD-3-Clause

[build-system]
build-backend = "rapids_build_backend.build"
requires = [
    "rapids-build-backend>=0.4.0,<0.5.0",
    "scikit-build-core[pyproject]>=0.11.0",
] # This list was generated by `rapids-dependency-file-generator`. To make changes, edit ../../dependencies.yaml and run `rapids-dependency-file-generator`.

[project]
name = "ucxx"
dynamic = ["version"]
description = "Python Bindings for the Unified Communication X library (UCX)"
readme = { file = "README.md", content-type = "text/markdown" }
authors = [
    { name = "NVIDIA Corporation" },
]
license = "BSD-3-Clause"
requires-python = ">=3.11"
dependencies = [
    "cuda-core>=0.3.2",
    "libucxx==0.49.*,>=0.0.0a0",
    "numpy>=1.23,<3.0",
    "nvidia-ml-py>=12",
    "rmm==26.4.*,>=0.0.0a0",
] # This list was generated by `rapids-dependency-file-generator`. To make changes, edit ../../dependencies.yaml and run `rapids-dependency-file-generator`.
classifiers = [
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Hardware",
    "Topic :: System :: Systems Administration",
    "Programming Language :: Python :: 3",
]

[project.entry-points."cmake.prefix"]
ucxx = "ucxx"

[project.optional-dependencies]
test = [
    "cloudpickle",
    "cudf==26.4.*,>=0.0.0a0",
    "cupy-cuda13x>=13.6.0",
    "numba-cuda>=0.22.1",
    "pytest",
    "pytest-asyncio>=1.0.0",
    "pytest-rerunfailures!=16.0.0",
    "rapids-dask-dependency==26.4.*,>=0.0.0a0",
] # This list was generated by `rapids-dependency-file-generator`. To make changes, edit ../../dependencies.yaml and run `rapids-dependency-file-generator`.

[project.urls]
Homepage = "https://github.com/rapidsai/ucxx"

[tool.rapids-build-backend]
build-backend = "scikit_build_core.build"
dependencies-file = "../../dependencies.yaml"
matrix-entry = "cuda_suffixed=true"
requires = [
    "cmake>=3.26.4,!=3.30.0",
    "cython>=3.2.2",
    "librmm==26.4.*,>=0.0.0a0",
    "libucxx==0.49.*,>=0.0.0a0",
    "ninja",
    "rmm==26.4.*,>=0.0.0a0",
] # This list was generated by `rapids-dependency-file-generator`. To make changes, edit ../../dependencies.yaml and run `rapids-dependency-file-generator`.

[tool.scikit-build]
build-dir = "build/{wheel_tag}"
cmake.build-type = "Release"
cmake.version = "CMakeLists.txt"
minimum-version = "build-system.requires"
ninja.make-fallback = false
sdist.exclude = ["*tests*"]
sdist.reproducible = true
wheel.packages = ["ucxx"]
wheel.exclude = ["*.pyx", "CMakeLists.txt"]

[tool.scikit-build.metadata.version]
provider = "scikit_build_core.metadata.regex"
input = "ucxx/VERSION"
regex = "(?P<value>.*)"

[tool.pydistcheck]
select = [
    "distro-too-large-compressed",
]

# PyPI hard limit is 1GiB, but try to keep these as small as possible
max_allowed_size_compressed = '10Mi'

[tool.pytest.ini_options]
asyncio_mode = "auto"
